Remove use of :textile Haml filter
It's stopped working in Haml 5.x, printing an error message telling you
to install haml-contrib. It's installed, but it's not working, which
breaks the "Using Haml" page in the demo site.

Not a great out-of-the-box experience!

Given that the `:textile` filter is unlikely to be in widespread use,
and Haml 6.0 is currently incompatible with all of Nesta, I've decided
not to dedicate time towards investigating and fixing this right now.

See gma/nesta#190.